% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf.R
\name{lf}
\alias{lf}
\title{lf}
\usage{
lf(
  type = "x",
  folder,
  pattern = "",
  remove = NULL,
  include.dirs = FALSE,
  ignore.case = FALSE
)
}
\arguments{
\item{folder}{Forder to list.}

\item{pattern}{File name pattern you want to save.}

\item{remove}{Path name pattern you want to remove.}

\item{include.dirs}{Whether include directory.}

\item{ignore.case}{Whether ignore case when matching pattern.}

\item{mode}{Running mode, rf|f|r|x.}
}
\value{
File paths.
}
\description{
Shortcut function for list.files
Shortcut function for list.files.
}
\examples{
lf("x", R.home()) 
lf("f", R.home()) 
lf("r", R.home()) 
lf("rf", R.home()) 
lf("rf", R.home(), pattern = "R$") 
lf("rf", R.home(), pattern = "R$", remove = c("base")) 
lf("rf", R.home(), pattern = "R$", remove = c("base"), ignore.case = TRUE) 
lf("x", R.home()) 
lf("x", R.home()) 
}
\author{
Jiahao Wang
}
